requirements.txt
django
django-braces
django-simple-history
django-js-reverse
django-crispy-forms
django-extensions
python-dateutil
model_mommy
Faker
unipath
pytz
arrow
django-tz-detect
mysqlclient
git+https://github.com/erikvw/django-crypto-fields@develop#egg=django_crypto_fields
git+https://github.com/erikvw/django-revision@develop#egg=django_revision
git+https://github.com/botswana-harvard/edc-appointment@develop#egg=edc-appointment
git+https://github.com/botswana-harvard/edc-consent@develop#egg=edc-consent
git+https://github.com/botswana-harvard/edc-constants@develop#egg=edc_constants
git+https://github.com/botswana-harvard/edc-device@develop#egg=edc-device
git+https://github.com/botswana-harvard/edc-example@develop#egg=edc-example
git+https://github.com/botswana-harvard/edc-identifier@develop#egg=edc-identifier
git+https://github.com/botswana-harvard/edc-lab@develop#egg=edc-lab
git+https://github.com/botswana-harvard/edc-metadata@develop#egg=edc-metadata
git+https://github.com/botswana-harvard/edc-protocol@develop#egg=edc_protocol
git+https://github.com/botswana-harvard/edc-registration@develop#egg=edc-registration
git+https://github.com/botswana-harvard/edc-rule-groups@develop#egg=edc-rule-groups
git+https://github.com/botswana-harvard/edc-timepoint@develop#egg=edc-timepoint
git+https://github.com/botswana-harvard/edc-visit-schedule@develop#egg=edc-visit-schedule
git+https://github.com/botswana-harvard/edc-visit-tracking@develop#egg=edc-visit-tracking
git+https://github.com/botswana-harvard/edc-off-study@develop#egg=edc-off-study
